(feat) Goes to merge page when clicking on a merge req entry
