+++ src/app/components/document/doc_mergeList_entry.jsx
@@ -2,10 +2,13 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { browserHistory } from 'react-router';
 
+import * as merge from './../../actions/mergeActions.jsx';
+
 export class DocMergeListEntry extends React.Component {
   constructor(props) {
     super(props);
     this.openMerge = this.openMerge.bind(this);
+    this.displayMergeRequest = this.displayMergeRequest.bind(this);
   }
 
   openMerge(e) {
@@ -14,10 +17,15 @@ export class DocMergeListEntry extends React.Component {
     console.log('opening merge req');
   }
 
+  displayMergeRequest(e) {
+    e.preventDefault();
+    this.props.dispatch(merge.displayMerge());
+  }
+
   render() {
     return (
       <div>
-        <h4 className="merge-list-title" onClick={this.openMerge}>{this.props.prInfo.collaboratorMessage}</h4>
+        <h4 className="merge-list-title" onClick={this.displayMergeRequest}>{this.props.prInfo.collaboratorMessage}</h4>
         <br/>
         <div>Request by: {this.props.prInfo.username}</div>
         <hr/>
